+++ collector-media3-exoplayer/src/main/java/com/bitmovin/analytics/media3/exoplayer/player/Media3ExoPlayerContext.kt
@@ -58,8 +58,6 @@ internal class Media3ExoPlayerContext(private val player: Player) : PlayerContex
             return false
         }
 
-    val playWhenReady get() = player.playWhenReady
-
     val getUriOfCurrentMedia get() = player.currentMediaItem?.localConfiguration?.uri
 
     override val playerVersion: String