+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - [Internal] Refactored theoplayer to avoid direct statemachine calls
+- [Internal] Started with refactoring of media3 player, to avoid statemachine calls
 
 ### Fixed
 - Added missing device information to ad samples